fix: Resolving merge updates
