+++ docs/src/content/docs/03-features/06-catalog/02-tui.mdx
@@ -197,13 +197,13 @@ This behavior is available by default as of 1.1, with no experiment required.
 </Aside>
 </Since>
 
-<Before version="1.0.4">
+<Before version="1.0.5">
 
 In a future release, the catalog UI will read tag information from the front-matter of component `README.md` files.
 
 </Before>
 
-<Since version="1.0.4">
+<Since version="1.0.5">
 
 Catalog authors can attach tags to a component by adding a `tags` key to the component's `README.md` front-matter. Tags appear as colored pills next to the component in the list view, and as a row above the rendered README in the detail view.
