--- docs/src/data/changelog/v1.1.0/catalog-redesign-enabled-by-default.mdx
@@ -1,10 +1,0 @@
----
-version: "v1.1.0"
-category: "new-features"
----
-
-#### Redesigned catalog
-
-The `terragrunt catalog` command now ships a redesigned TUI that discovers modules, units, stacks, and boilerplate templates across the whole repository, presents them in `All`, `Modules`, and `Templates` tabs, and scaffolds a selected component through an interactive form that prompts for each variable it exposes. Authors can keep directories out of discovery with a `.terragrunt-catalog-ignore` file.
-
-The redesigned catalog is enabled by default. Previously gated behind the `catalog-redesign` experiment, it no longer requires `--experiment catalog-redesign`.
